requirements.txt
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.24.0
fastmcp>=0.1.0
